pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "agency-swarm"
dynamic = ["version"]
authors = [{ name = "VRSEN", email = "me@vrsen.ai" }]
description = "An open source agent orchestration framework built on top of the latest OpenAI Responses API."
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "datamodel-code-generator==0.28.3",
    "deepdiff==6.7.1",
    "docstring_parser==0.16",
    "jsonref==1.1.0",
    "openai>=1.66.0,<2.0.0",
    "pydantic==2.8.2",
    "python-dotenv==1.0.1",
    "rich==13.7.1",
    "termcolor>=2.3.0",
]
requires-python = ">=3.10"
urls = { homepage = "https://github.com/VRSEN/agency-swarm" }

[project.scripts]
agency-swarm = "agency_swarm.cli:main"

[tool.setuptools_scm]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
